Invoke Alliance of 1373 is now UK NF, not decision
Added the Focus which directly adds Portugal to UKs faction on completion, instead of a event where Portugal could say no. Since the Focus isn't repeatable compared to the decision.
Made the Focus mutually exclusive to Coerce Spain, but not Fortify the Rock.
Slightly moved some foci around to make space for the new one.