Removed Send Volunteer ai_strategy from Germany and Italy
Somewhat a shot in a dark but when I had the error log open in our MP game I saw the error: 
[18:38:52][persistent.cpp:52]: Error: "Malformed token: 19735, near line: 123244" in file: "" near line: 123244

the malformed token 19735 I saw multiple times while the game was running.

The token didn't appear in our current or last week savegame but in the one from 2020/11/09 21:14, so while the game was running.

In this savegame it appears 2 times in ai_strategy. One for Germany one for Italy.
			ai_strategy={ type=34 id=19735 value=1000 }
type=34 ai_strategy is 	send_volunteers_desire strategy says the HoI IV wiki.
So I just removed the send_volunteers_desire ai_strategy from Germany and Italy, potentially this is another ai_strategy not beeing able to be used in ai_strategy_plans, without saying so in the error.log or debug. Doesn't hurt that much to remove it.