[N] Removed Italy First NF Branch under IHMP historical game rule
Removed the "Italy First" Focus Branch if game_rule historical is set.
For now it also removes wargoal on greece which is shared with pact of steel, but since the group disallows Italy taking the focus before WW2 started either way, and after WW2 it is always cheaper to just fabricate manually, I think this is fine. Makes it clear to new players as well, since this was never codified in the rules.