[F] Removed randomness which warlords willingly becomes a puppet of China, if China uses subjugate warlord focus
To have an interesting civil war setting I removed the randomness which warlords resists and which warlord willingly becomes a puppet of China. Also staying true to history (Yunnan partnering with China, Guanxi and Shanxi resisting, Xinjiang far away from anything and communist)
This allows PRC to conquer Shanxi for a better fighting base, and China to attack Guanxi Clique for Army XP to remove corruption.